+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Suporte a criação de Alias com Condição
 * Melhoria preenchimentos objetos relacionais
 * ORWHERE ou ORHAVING em Blocos
+* DataTypes Float e Decimal com round nas casas decimais
 
 ### 0.0.4
 * Correção Comparação de Objetos e PK's Diferentes